Reorganize assets
